First normalization and test hook

+++ simplifications/anonymous_1984/scratch/anonymous_compiling.c
@@ -1,3 +1,3 @@
-void read(int j, char *i, int p){write(j/p+p,i---j,(int)i/(int)i);}
+void read(int j, char *i, int p){write(j/p+p,i---j,i/i);}
 int i;main(){for(;i["]<i;++i){--i;}"];read('-'-'-',i+++"hell\
 o, world!\n",'/'/'/'));}
